feat: create security test files for services
