+++ src/services/projectService.security.spec.ts
@@ -1,0 +1,8 @@
+import { PrismaClient } from "@prisma/client";
+const faker = require("@faker-js/faker");
+
+const prisma = new PrismaClient();
+
+afterAll(() => {
+  return expect(prisma.$disconnect()).resolves;
+});
